Move StreakUtil to logic module
